requirements.txt
# Google Cloud Platform (Python 3.12 compatible versions)
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-cloud-error-reporting>=1.10.0
google-api-python-client>=2.100.0

# Apache Airflow (Python 3.12 compatible)
apache-airflow>=2.8.0
apache-airflow-providers-google>=10.11.0
apache-airflow-providers-http>=4.8.0

# dbt (data build tool) - Python 3.12 compatible
dbt-core>=1.7.0
dbt-bigquery>=1.7.0

# Data scraping and processing
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
pandas>=2.0.0

# Environment and config
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Testing
pytest>=8.0.0
